+++ src/main/java/com/turnosrotativos/service/IConceptoLaboralService.java
@@ -5,5 +5,5 @@ import com.turnosrotativos.entity.ConceptoLaboral;
 import java.util.List;
 
 public interface IConceptoLaboralService {
-    List<ConceptoLaboral> obtenerConceptosLaborales(Integer id, String nombre);
+    List<ConceptoLaboral> obtenerConceptosLaborales(Long id, String nombre);
 }
